Agregados colores para las paletas. cambiado archivo de creación
