+++ src/assets/constantes.js
@@ -1,18 +1,15 @@
 const plantillas = [
 	{
 		name: 'Formulario en blanco',
-		imageSrc: 'https://tailwindui.com/img/ecommerce-images/home-page-02-edition-01.jpg',
-		imageAlt:
-			'Desk with leather desk pad, walnut desk organizer, wireless keyboard and mouse, and porcelain mug.',
-		href: '#',
+		colores: ['#e2e2e2', '#76777c', '#141a35']
 	},
 	{
 		name: 'Paleta Rosa',
-		imageSrc: 'https://tailwindui.com/img/ecommerce-images/home-page-02-edition-02.jpg',
+		colores: ['#0097ff', '#cb98ff', '#ffb3ff'],
 	},
 	{
 		name: 'Paleta Azul',
-		imageSrc: 'https://tailwindui.com/img/ecommerce-images/home-page-02-edition-03.jpg',
+		colores: ['#00c6cf', '#0076ab', '#00318c'],
 	},
 ];
 export { plantillas };
